requirements.txt
# Dependências do projeto
python-telegram-bot
requests
sqlite3
